feature(element): adición de toast y alerts
Adición de los toast y alerts para la implementación de salas, boton de editar en View, boton de ver en editar y cambios en el sidebar.

+++ reservas-asw-front/src/app/admin/pages/rooms/list/list-room.component.ts
@@ -2,6 +2,8 @@ import { Component } from '@angular/core';
 import { Room, RoomResponse , RoomsResponse } from '../../../interfaces/rooms.interfaces';
 import { RouteName } from '../../../../../utils/enums';
 import { RoomsService } from '../../../services/rooms.service';
+import { AlertsService } from 'src/app/services/alerts.service';
+import { ToastsService } from 'src/app/services/toasts.service';
 
 @Component({
   selector: 'app-list-room',
@@ -13,7 +15,9 @@ export class ListRoomComponent {
   rooms: Room[] = [];
 
   constructor(
-    private roomsService: RoomsService
+    private roomsService: RoomsService,
+    private toastService: ToastsService,
+    private alertsService: AlertsService
   ) { }
 
   ngOnInit(): void {
@@ -28,9 +32,23 @@ export class ListRoomComponent {
   }
 
   deleteRoom( id: number ): void {
-    this.roomsService.deleteRoom(id)
-        .subscribe(
-          (roomResponse: RoomResponse) => this.rooms = this.rooms.filter((room: Room) => room.idSala !== id)
-        );
+    this.alertsService.showConfirmDialog({
+      message: '¿Desea eliminar la sala, esta acción no se podrá revertir?',
+      header: 'Eliminar sala',
+    })
+      .then(resp => {
+        if (resp) {
+          this.roomsService.deleteRoom(id)
+            .subscribe(
+              (roomResponse: RoomResponse) => {
+                this.rooms = this.rooms.filter((room: Room) => room.idSala !== id);
+                this.toastService.showToastSuccess({ summary: 'Sala eliminada', detail: 'La sala ha sido eliminada correctamente.' });
+              }
+            );
+        } else {
+          return;
+        }
+      })
+      .catch(console.log);
   }
 }
